remove deprecated BaselineSessionEventsListenerBuilderFactoryFactory
